Minor Update: Separação de Responsabilidades em GenericListTab; Correção de bugs (ex.: Overflow em BottomNavigationBar).
