+++ lib/features/data/models/generic_list_item_model.dart
@@ -4,7 +4,6 @@ class GenericListItemModel {
   final String? lowerHeaderField;
   final int isActive;
   final DateTime? initialDate;
-  final DateTime? finalDate;
   final DateTime? lastUpdatedAt;
 
   GenericListItemModel({
@@ -13,7 +12,6 @@ class GenericListItemModel {
     required this.lowerHeaderField,
     required this.isActive,
     this.initialDate,
-    this.finalDate,
     this.lastUpdatedAt,
   });
 }
