+++ lib/features/modules/panel/panel_page.dart
@@ -65,26 +65,36 @@ class _PanelPageState extends State<PanelPage> {
           tabName: 'Inventories',
           items: inventories,
           searchParameters: 'Título ou Id',
+          firstDetailFieldName: 'Criado em',
+          secondDetailFieldName: 'Última atualização em',
         ),
         GenericListTab(
           tabName: 'Domains',
           items: domains,
           searchParameters: 'Título ou Id',
+          firstDetailFieldName: 'Criado em',
+          secondDetailFieldName: 'Última atualização em',
         ),
         GenericListTab(
           tabName: 'Tags',
           items: tags,
           searchParameters: 'Serial ou Id',
+          firstDetailFieldName: 'Criado em',
+          secondDetailFieldName: 'Visto pela última vez em',
         ),
         GenericListTab(
           tabName: 'Readers',
           items: readers,
           searchParameters: 'Nome ou MAC',
+          firstDetailFieldName: 'Criado em',
+          secondDetailFieldName: 'Visto pela última vez em',
         ),
         GenericListTab(
           tabName: 'Members',
           items: members,
           searchParameters: 'Nome ou Email',
+          firstDetailFieldName: 'Criado em',
+          secondDetailFieldName: 'Visto pela última vez em',
         ),
         const EntitiesTab(),
         const AdminTab(),
