+++ lib/features/modules/panel/widgets/generic_list_tab.dart
@@ -1,9 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:infinite_scroll_pagination/infinite_scroll_pagination.dart';
-import 'package:intl/intl.dart';
 import 'package:inventory_platform/features/data/models/generic_list_item_model.dart';
 import 'package:inventory_platform/features/data/models/organization_model.dart';
+import 'package:inventory_platform/features/modules/panel/widgets/generic_list_header.dart';
+import 'package:inventory_platform/features/modules/panel/widgets/generic_list_item_card.dart';
+import 'package:inventory_platform/features/modules/panel/widgets/search_bar_widget.dart';
 import 'package:inventory_platform/features/widgets/custom_progress_indicator.dart';
 import 'package:inventory_platform/features/widgets/list_item_skeleton.dart';
 import 'package:inventory_platform/features/widgets/temporary_message_display.dart';
@@ -13,10 +15,14 @@ class GenericListTab extends StatefulWidget {
   final List<GenericListItemModel> items;
   final String tabName;
   final String? searchParameters;
+  final String? firstDetailFieldName;
+  final String? secondDetailFieldName;
 
   const GenericListTab({
     super.key,
     this.searchParameters,
+    this.firstDetailFieldName,
+    this.secondDetailFieldName,
     required this.tabName,
     required this.items,
   });
@@ -26,7 +32,7 @@ class GenericListTab extends StatefulWidget {
 }
 
 class _GenericListTabState extends State<GenericListTab> {
-  final OrganizationModel organization = Get.arguments;
+  final OrganizationModel _organization = Get.arguments;
   final TextEditingController _searchController = TextEditingController();
   final PagingController<int, GenericListItemModel> _pagingController =
       PagingController(firstPageKey: 0);
@@ -38,7 +44,7 @@ class _GenericListTabState extends State<GenericListTab> {
   @override
   void initState() {
     _pagingController.addPageRequestListener((pageKey) => _fetchPage(pageKey));
-    _searchController.addListener(_filterInventories);
+    _searchController.addListener(_filterItems);
     _allItemsList = widget.items
       ..sort((a, b) => b.initialDate!.compareTo(a.initialDate!));
     super.initState();
@@ -54,9 +60,16 @@ class _GenericListTabState extends State<GenericListTab> {
       child: Column(
         crossAxisAlignment: CrossAxisAlignment.start,
         children: [
-          _buildHeader(organization.title),
-          _buildSearchBar(),
-          _buildInventoryList(),
+          GenericListHeader(
+            tabName: widget.tabName,
+            itemCount: _allItemsList.length,
+            organizationName: _organization.title,
+          ),
+          SearchBarWidget(
+            searchController: _searchController,
+            hintText: 'Pesquisar por ${widget.searchParameters ?? '...'}',
+          ),
+          _buildList(),
         ],
       ),
     );
@@ -82,7 +95,7 @@ class _GenericListTabState extends State<GenericListTab> {
     final startIndex = pageKey * pageSize;
     final endIndex = startIndex + pageSize;
 
-    final paginatedInventories = _filteredItemsList.sublist(
+    final paginatedItems = _filteredItemsList.sublist(
       startIndex,
       endIndex > _filteredItemsList.length
           ? _filteredItemsList.length
@@ -93,19 +106,19 @@ class _GenericListTabState extends State<GenericListTab> {
 
     if (mounted) {
       if (isLastPage) {
-        _pagingController.appendLastPage(paginatedInventories);
+        _pagingController.appendLastPage(paginatedItems);
       } else {
-        _pagingController.appendPage(paginatedInventories, pageKey + 1);
+        _pagingController.appendPage(paginatedItems, pageKey + 1);
       }
     }
   }
 
-  void _filterInventories() {
+  void _filterItems() {
     final query = _searchController.text.toLowerCase();
     setState(() {
-      _filteredItemsList = _allItemsList.where((inventory) {
-        return inventory.upperHeaderField.toLowerCase().contains(query) ||
-            (inventory.id ?? ' ').toLowerCase().contains(query);
+      _filteredItemsList = _allItemsList.where((item) {
+        return item.upperHeaderField.toLowerCase().contains(query) ||
+            (item.id ?? ' ').toLowerCase().contains(query);
       }).toList();
     });
     _pagingController.itemList?.clear();
@@ -116,188 +129,20 @@ class _GenericListTabState extends State<GenericListTab> {
     if (mounted) _pagingController.refresh();
   }
 
-  Widget _buildHeader(String organizationName) {
-    return Padding(
-      padding: const EdgeInsets.only(
-          left: 16.0, right: 16.0, top: 20.0, bottom: 8.0),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Row(
-            crossAxisAlignment: CrossAxisAlignment.center,
-            children: [
-              Text(
-                widget.tabName,
-                style: const TextStyle(
-                  fontSize: 24.0,
-                  fontWeight: FontWeight.bold,
-                  color: Colors.black87,
-                ),
-              ),
-              Padding(
-                padding: const EdgeInsets.symmetric(horizontal: 8.0),
-                child: Container(
-                  padding: const EdgeInsets.symmetric(horizontal: 8.0),
-                  decoration: const BoxDecoration(
-                    color: Color.fromARGB(121, 158, 158, 158),
-                    borderRadius: BorderRadius.all(Radius.circular(10.0)),
-                  ),
-                  child: Text(
-                    '${_allItemsList.length}',
-                    style: const TextStyle(
-                      fontSize: 16.0,
-                      fontWeight: FontWeight.bold,
-                      color: Colors.black87,
-                    ),
-                  ),
-                ),
-              ),
-            ],
-          ),
-          Padding(
-            padding: const EdgeInsets.only(top: 4.0),
-            child: Container(
-              padding: const EdgeInsets.symmetric(
-                horizontal: 8.0,
-                vertical: 4.0,
-              ),
-              decoration: BoxDecoration(
-                color: Colors.greenAccent.shade700,
-                borderRadius: BorderRadius.circular(5.0),
-              ),
-              child: Text(
-                organizationName,
-                style: const TextStyle(
-                  color: Colors.white,
-                  fontWeight: FontWeight.bold,
-                ),
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildSearchBar() {
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 16.0, vertical: 8.0),
-      child: Container(
-        decoration: BoxDecoration(
-          color: Colors.grey.shade300,
-          borderRadius: BorderRadius.circular(12.0),
-          boxShadow: [
-            BoxShadow(
-              color: Colors.black.withOpacity(0.05),
-              blurRadius: 10.0,
-              offset: const Offset(0, 4),
-            ),
-          ],
-        ),
-        child: TextField(
-          controller: _searchController,
-          style: const TextStyle(fontSize: 16.0),
-          cursorColor: Colors.black87,
-          decoration: InputDecoration(
-            contentPadding: const EdgeInsets.symmetric(vertical: 12.0),
-            hintText: 'Pesquisar por ${widget.searchParameters ?? '...'}',
-            hintStyle: TextStyle(
-              color: Colors.grey.shade700,
-              fontSize: 16.0,
-            ),
-            prefixIcon: const Icon(
-              Icons.search,
-              color: Colors.black54,
-            ),
-            filled: true,
-            fillColor: Colors.transparent,
-            border: OutlineInputBorder(
-              borderRadius: BorderRadius.circular(12.0),
-              borderSide: BorderSide.none,
-            ),
-            enabledBorder: OutlineInputBorder(
-              borderRadius: BorderRadius.circular(12.0),
-              borderSide: BorderSide.none,
-            ),
-            focusedBorder: OutlineInputBorder(
-              borderRadius: BorderRadius.circular(12.0),
-              borderSide: BorderSide(
-                color: Colors.black.withOpacity(0.2),
-                width: 1.5,
-              ),
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-
-  Widget _buildInventoryCard(GenericListItemModel item) {
-    final formattedOpenedAt = formatDate(item.initialDate);
-    final formattedLastUpdatedAt = formatDate(item.lastUpdatedAt);
-
-    return Card(
-      margin: const EdgeInsets.symmetric(vertical: 8.0),
-      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(12.0)),
-      elevation: 4.0,
-      color: Colors.white,
-      child: ListTile(
-        contentPadding:
-            const EdgeInsets.symmetric(vertical: 12.0, horizontal: 16.0),
-        title: Text(item.upperHeaderField,
-            style: const TextStyle(
-                fontWeight: FontWeight.bold,
-                fontSize: 16.0,
-                color: Colors.black87)),
-        subtitle: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            if (item.lowerHeaderField != null)
-              Text(item.lowerHeaderField!,
-                  style: TextStyle(color: Colors.grey.shade700)),
-            const SizedBox(height: 4),
-            Text('Aberto em: $formattedOpenedAt',
-                style: TextStyle(color: Colors.grey.shade600, fontSize: 12.0)),
-            Text('Atualizado em: $formattedLastUpdatedAt',
-                style: TextStyle(color: Colors.grey.shade600, fontSize: 12.0))
-          ],
-        ),
-        trailing: Stack(
-          alignment: Alignment.center,
-          children: [
-            Icon(
-              item.isActive == 1
-                  ? Icons.circle_outlined
-                  : Icons.circle_outlined,
-              color: item.isActive == 1 ? Colors.green : Colors.red,
-              size: 20.0,
-            ),
-            Icon(
-              item.isActive == 1 ? Icons.circle : Icons.circle,
-              color: item.isActive == 1 ? Colors.green : Colors.red,
-              size: 12.0,
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  String formatDate(DateTime? date) {
-    return date != null
-        ? DateFormat.yMMMMd().format(date)
-        : "Data Indisponível";
-  }
-
-  Widget _buildInventoryList() {
+  Widget _buildList() {
     return Expanded(
       child: PagedListView<int, GenericListItemModel>(
         padding: const EdgeInsets.only(
             left: 16.0, right: 16.0, top: 16.0, bottom: 32.0),
         pagingController: _pagingController,
         builderDelegate: PagedChildBuilderDelegate<GenericListItemModel>(
-          itemBuilder: (context, inventory, index) {
-            return _buildInventoryCard(inventory);
+          itemBuilder: (context, item, index) {
+            return GenericListItemCard(
+              item: item,
+              firstDetailFieldName: widget.firstDetailFieldName ?? '',
+              secondDetailFieldName: widget.secondDetailFieldName ?? '',
+              key: ValueKey(item.id),
+            );
           },
           firstPageProgressIndicatorBuilder: (_) =>
               const CustomProgressIndicator(),
