+++ lib/features/modules/panel/widgets/search_bar_widget.dart
@@ -1,0 +1,66 @@
+import 'package:flutter/material.dart';
+
+class SearchBarWidget extends StatelessWidget {
+  final TextEditingController searchController;
+  final String hintText;
+
+  const SearchBarWidget({
+    super.key,
+    required this.searchController,
+    required this.hintText,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.symmetric(horizontal: 16.0, vertical: 8.0),
+      child: Container(
+        decoration: BoxDecoration(
+          color: Colors.grey.shade300,
+          borderRadius: BorderRadius.circular(12.0),
+          boxShadow: [
+            BoxShadow(
+              color: Colors.black.withOpacity(0.05),
+              blurRadius: 10.0,
+              offset: const Offset(0, 4),
+            ),
+          ],
+        ),
+        child: TextField(
+          controller: searchController,
+          style: const TextStyle(fontSize: 16.0),
+          cursorColor: Colors.black87,
+          decoration: InputDecoration(
+            contentPadding: const EdgeInsets.symmetric(vertical: 12.0),
+            hintText: hintText,
+            hintStyle: TextStyle(
+              color: Colors.grey.shade700,
+              fontSize: 16.0,
+            ),
+            prefixIcon: const Icon(
+              Icons.search,
+              color: Colors.black54,
+            ),
+            filled: true,
+            fillColor: Colors.transparent,
+            border: OutlineInputBorder(
+              borderRadius: BorderRadius.circular(12.0),
+              borderSide: BorderSide.none,
+            ),
+            enabledBorder: OutlineInputBorder(
+              borderRadius: BorderRadius.circular(12.0),
+              borderSide: BorderSide.none,
+            ),
+            focusedBorder: OutlineInputBorder(
+              borderRadius: BorderRadius.circular(12.0),
+              borderSide: BorderSide(
+                color: Colors.black.withOpacity(0.2),
+                width: 1.5,
+              ),
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+}
